test(e2e): SMI-4970 runtime staging-skill discovery + path #5 401 tolerance
The usage-counter E2E tests hard-coded a staging skill ID that rotted
twice on seed-data drift (anthropic/commit, then
anthropics/web-artifacts-builder). Replace it with runtime discovery:
resolveStagingSkillId() reads the staging skills table directly via
PostgREST with the service-role key (not the rate-limited skills-search
edge function), memoized, with a timeout + 1 retry and distinct
fail-fast messages for empty-result vs unreachable. SKILLSMITH_E2E_SKILL_ID
remains as an escape hatch. Both cli + mcp suites resolve the skill in
beforeAll, de-duplicating the constant.

account-usage path #5 asserted [200,429]; staging anonymous-trial
exhaustion returns 401 — add it to the tolerated set (the negative
counter assertion holds for 200/429/401 alike).

The added helper pushed usage-counter-fixture.ts past the 500-line
limit; split it into usage-counter-fixture.{types,helpers}.ts
(behavior-preserving code movement; all public exports preserved).
